usage: show all command aliases.
Such that when building command usage synopsis each command name is
listed along with the automatic usage string.

Initially (and much easier is install|ins|i) but this seems confusing in
the synopsis so the decision was made to have some repetition (ie, in
the automatic option string) but increase legibility of command aliases
(names).