Support converters as object maps.
This allows a converter function to be an object, the keys of the object
are mapped to command keys allowing validation/coercion for an argument
to be mutated depending upon the supplied command.

The object should contain a *default* property which will be used when
a) no command is specified or b) the supplied command alias does not map
to a property of the converter object.

Update default *trace* configuration boolean to whether NODE_ENV is
test.