Support error interception.
If a program declares an *intercept* function in the error configuration
it is invoked when an error occurs with the wrapped error, req and the
source error parameters (err, parameters, e) if it returns true then the
error is raised effectively deferring to the default error flow.

If it returns false then flow control is deferred to the error intercept
method.

This allows program that wish to modify the error printing logic based
upon configuration to work as expected, for example, under some
circumstances a program may wish to print all errors as JSON.

Note that if the intercept method is handling the error, the *raise*
event will not fire, if the program is listening for raise the intercept
method should emit it.