Add notify middleware.
Whilst the events middleware is useful for basic logic it is not
asynchronous and is early in the middleware processing.

The notify middleware allows event based notifications for options using
a *notify:* event name prefix.

These event listeners are passed a next function and may execute
asynchronously. By default this is placed late in the flow so that rc
files, environment variables are available in the req and option
properties are already set on the stash.

Also allow an error to set *bail* to *true* to halt middleware
processing whilst maintaining the deprecated support for an error being
=== true.