Collect request errors into req.errors object.
And add a cause property when the middleware processing was aborted or
bailed.

The req.errors object has a *list* property containing all errors
encountered (except internal cause errors) and a *has()* function to
determine if any errors occurred during processing.

Useful if you need to test for an error condition in a complete listener
or parse callback.