allow help middleware to be configured via configure()
I'm unsure why `help` is unable to be configured in this manner, unlike many of the other middlewares.  It seems `help` was not originally middleware?

There's a conditional regarding the description.  Current behavior is to grab the description out of the `Command` instance if `help.description = true` in the config.  So, we check for that value, and if it's present, we pull the description out of the instance.

I moved a couple default strings into the configuration defaults as well.